Extract KMeans TClusters into a separate file
